third_party/blink/web_tests/external/wpt/css/cssom/HTMLLinkElement-disabled-007.tentative-expected.txt

This is a testharness.js-based test.
FAIL HTMLLinkElement.disabled setter sets the explicitly enabled state if toggled back and forth. assert_equals: expected "rgb(0, 128, 0)" but got "rgba(0, 0, 0, 0)"
Harness: the test ran to completion.
